fix: integer type on collect_database method

+++ src/database/read.c
@@ -312,7 +312,7 @@ static size_t collect_database(struct Database **database, const int fd, char *b
 
   struct KVPair **cache = (*database)->data;
 
-  for (uint32_t i = 0; i < *count; ++i) {
+  for (uint64_t i = 0; i < *count; ++i) {
     struct KVPair *kv = malloc(sizeof(struct KVPair));
     collected_bytes += collect_kv(kv, fd, block, block_size, at);
